requirements.txt
# Scraping
Telethon==1.34.0
python-dotenv==1.0.1
pandas==2.2.0

# API
fastapi==0.109.0
uvicorn==0.27.0
pydantic==2.6.0

# Database
psycopg2-binary==2.9.9
sqlalchemy==2.0.25

# Data Engineering
dagster==1.6.3
dagster-webserver==1.6.3
dbt-core==1.7.0
dbt-postgres==1.7.0

# Computer Vision
ultralytics==8.1.14
opencv-python-headless==4.9.0.80

# Utils
tqdm==4.66.1
